choose_disk: initial commit
When ceph-ansible needs to choose disks on hosts it currently use the short
disk path like /dev/sdb,/dev/sdc, ...

This naming is a big issue as :
- its not consistent over time : adding|removing a disk or just rebooting can change the path
- it doesn't insure the features we are looking for : Does /dev/sdb is the 2TB
  disk we are looking for ? Does /dev/sdc is my SSD ?
- Declaring multiple devices is pretty verbose : Selecting 5 disks means adding
  5 /dev/sd* entry

Using /dev/disk/by-id/ paths instead of /dev/sd* was a first idea but prooved to
be wrong as :
- the path usually includes the serial number of the device making the disk listing long & painful
- some hosts provides a pretty unreadable name like scsi-36848f690e68a50001bbef0a0094625c1

So we need a different approach that is
- easier to use (no need to know the device path)
- dense (adding similar devices type should make a short list)
- descriptive (select devices by their features)
- consistent (a reboot or a disk change should not affect the result)

This patch is adding a new description language to select disk's devices.
It also includes the associated tests to perform the unitary tests

This first commit introduce the main idea and a basic implementation.
Following features & improvment will be added by further commits

A Python Ansible module is added and called as a new task like :
tasks:
    # If we can't get python2 installed before any module is used we will fail
    # so just try what we can to get it installed
    - name: gathering facts
      setup:
    - name: choose disk
      choose_disk:
        facts: "{{ansible_devices}}"
        disks: "{{lookup_disks}}"

The ansible_devices is provided by Ansible once the setup call is performed : it can be manually run by
 [root@host]: ansible localhost -m setup -a filter="ansible_devices"

The lookup_disks variable is adjusted by users to filter the devices they are interested in using a syntax like :
    {'storage_name': { 'feature_1': 'value_1', 'feature_2': 'value_2', 'count': number_of_devices_we_search }}

A typical example is :
    vars:
      lookup_disks: "{'storage_disks': {'model': 'SAMSUNG MZ7LN512', 'rotational': '1', 'count': 2 }}"

This sample description reports we are looking for 2 HDDs of that M27LN512 type.

All the feature names & values are the one reported in 'ansible_devices' except
the "count" which is a specific one of our own to indicate the number of devices
of that type we want.

As a result, a new fact is populated with the associated devices.

Signed-off-by: Erwan Velu <erwan@redhat.com>