choose_disk: Avoid globals() for handling operators
When detecting operators in a match, the associated function was assumed
to be the name of the function prefixed with underscore.

That was requiring to use the globals() which isn't that nice.
More than that, it would allow to run any function that is starting by
an underscore which isn't that nice.

This patch reimplement that part by creating an association between
user's keywords and internal functions. This way its no more possible to
run any other functions.

Signed-off-by: Erwan Velu <erwan@redhat.com>