choose_disk: Disks can be data + journal
The existing code made the assomption that a disk can be data or
journal, while it can be both at the same time (as per its sub
partitions).

First, this patch fix the sorting of the partitions to be sure getting
them always in the same order and get so a consistent parsing.

Then, the type (data | journal) is added to the 'ceph_prepared'
attribute.

Signed-off-by: Erwan Velu <erwan@redhat.com>