choose_disk: Reporting logger when reaching fail_json()
When using choose_disk in ansible, when fatal() is called, only the
message is reported leading to something like :

    Could only find 0 of the 3 expected devices

This is clearly not enough to understand 'why' it failed.

One option is to setup the logging into a file, which is useful for
debugging but make it more difficult to use in real-life setups.

This patch is about logging into memory and pad all the logger messages
in the message returned to ansible. This way, all the debugging is
reachable from the ansible error.

Signed-off-by: Erwan Velu <erwan@redhat.com>