choose_disk: Refusing partions in legacy mode
When users are choosing the legacy mode, they have to pass a full disk.
If a partition is given, an error should be raised to the user.

This commit is introducing getParentBlock() which reports the name of
the parent device.
- If nothing is returned, it means that's already the full block device.
- If something is returned, it features the name of the parent device

The lsblk command is coming from Sebastien Han. Thx Bro !

Signed-off-by: Erwan Velu <erwan@redhat.com>