choose_disk: Handle osd_objectstore in legacy mode
When osd_objectstore is set to bluestore, the legacy output should be different.

The 'devices' shall be reported in the "block" output.
The 'dedicated_devices' shall be reported in the "block.wal" output.

This commit is adjusting this behavior based on the value of
osd_objectstore extracted from the ansible playbook.

Signed-off-by: Erwan Velu <erwan@redhat.com>